rename: 프로필 패키지명 변경
