+++ mcnc_2024/src/router/home/user-profile/user-profile.js
@@ -9,14 +9,14 @@ const profileRoutes = [
     {
         path : "/profile",
         name : "Profile",
-        component : () => import("@/components/home/user_profile/UserProfile.vue"),
+        component : () => import("@/components/home/user-profile/UserProfile.vue"),
         children : [
         ]
     },
     {
         path : "/profile/update",
         name : "UpdateProfile",
-        component : () => import("@/components/home/user_profile/profile_list/UpdateProfile.vue"),
+        component : () => import("@/components/home/user-profile/profile-menu/UpdateProfile.vue"),
     }
 ]
